refactoring of obstacles rendering and grid pipeline plotObstacles->renderObstacles/buildObstacleGrid instead of everything in showObstacles()
