+++ wos-interactive-map-front/src/three/obstacles/obstaclesController.js
